# docs/api-documentation-openapi.json
{
    "openapi": "3.0.1",
...
    "info": {
        "title": "mimoto",
        "description": "mimoto",
        "license": {
            "name": "Mosip",
            "url": "https://docs.mosip.io/platform/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://api.collab.mosip.net/v1/mimoto",
            "description": "For Swagger"
        }
    ],
    "paths": {
...
        "/issuers": {
            "get": {
                "operationId": "getAllIssuers",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "tags": [
                    "Trusted Issuers"
                ],
                "description": "This API provides data with search capability to populate the list of supported issuers in Inji Web, which is then displayed under the List of Issuers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "issuers": [
                                            {
                                                "credential_issuer": "string",
                                                "protocol": "string",
                                                "display": [
                                                    {
                                                        "name": "string",
                                                        "logo": {
                                                            "url": "string",
                                                            "alt_text": "string"
                                                        },
                                                        "title": "string",
                                                        "description": "string",
                                                        "language": "string"
                                                    }
                                                ],
                                                "client_id": "string",
                                                "redirect_uri": "string",
                                                "scopes_supported": [
                                                    "string"
                                                ],
                                                "authorization_endpoint": "string",
                                                "authorization_audience": "string",
                                                "token_endpoint": "string",
                                                "proxy_token_endpoint": "string",
                                                "credential_endpoint": "string",
                                                "credential_type": [
                                                    "string"
                                                ],
                                                "credential_audience": "string",
                                                "client_alias": "string",
                                                "additional_headers": {
                                                    "additionalProp1": "string",
                                                    "additionalProp2": "string",
                                                    "additionalProp3": "string"
                                                },
                                                "wellknown_endpoint": "string"
                                            }
                                        ]
                                    }
                                },
                                "x-examples": {
                                    "Example 1": {
                                        "id": "mosip.resident.vid",
                                        "version": "v1",
                                        "str": null,
                                        "responsetime": "2022-10-31T05:08:14.846Z",
                                        "metadata": null,
                                        "response": {
                                            "issuers": [
                                                {
                                                    "credential_issuer": "MOSIPInsurance",
                                                    "protocol": "OpenId4VCI",
                                                    "display": [
                                                        {
                                                            "name": "MOSIP Insurance",
                                                            "logo": {
                                                                "url": "https://api.collab.mosip.net/inji/mosip-logo.png",
                                                                "alt_text": "MOSIP-Logo"
                                                            },
                                                            "title": "Download via LIC",
                                                            "description": "This API provides data to populate the list of supported issuers in Inji Web, which is then displayed under the List of Issuers section on the Home page of Inji Web.",
                                                            "language": "en"
                                                        }
                                                    ],
                                                    "client_id": "3yz7-j3xRzU3SODdoNgSGvO_cD8UijH3AIWRDAg1x-M",
                                                    ".well-known": "http://localhost:8088/.well-known/openid-credential-issuer"
                                                },
                                                {
                                                    "credential_issuer": "MOSIPNationalID",
                                                    "protocol": "OpenId4VCI",
                                                    "display": [
                                                        {
                                                            "name": "MOSIP National ID",
                                                            "logo": {
                                                                "url": "https://api.collab.mosip.net/inji/mosip-logo.png",
                                                                "alt_text": "MOSIP-Logo"
                                                            },
                                                            "title": "Download via LIC",
                                                            "description": "Enter your policy number to download your card.",
                                                            "language": "en"
                                                        }
                                                    ],
                                                    "client_id": "3yz7-j3xRzU3SODdoNgSGvO_cD8UijH3AIWRDAg1x-M",
                                                    ".well-known": "http://localhost:8088/.well-known/openid-credential-issuer"
                                                }
                                            ]
                                        }
                                    },
                                    "errors": null
                                },
                                "examples": {
                                    "Success response": {
                                        "value": {
                                            "id": "mosip.resident.vid",
                                            "version": "v1",
                                            "str": null,
                                            "responsetime": "2022-10-31T05:08:14.846Z",
                                            "metadata": null,
                                            "response": {
                                                "issuers": [
                                                    {
                                                        "credential_issuer": "MOSIPInsurance",
                                                        "protocol": "OpenId4VCI",
                                                        "display": [
                                                            {
                                                                "name": "MOSIP Insurance",
                                                                "logo": {
                                                                    "url": "https://api.collab.mosip.net/inji/mosip-logo.png",
                                                                    "alt_text": "MOSIP-Logo"
                                                                },
                                                                "title": "Download via LIC",
                                                                "description": "Enter your policy number to download your card.",
                                                                "language": "en"
                                                            }
                                                        ],
                                                        "client_id": "3yz7-j3xRzU3SODdoNgSGvO_cD8UijH3AIWRDAg1x-M",
                                                        ".well-known": "http://localhost:8088/.well-known/openid-credential-issuer"
                                                    },
                                                    {
                                                        "credential_issuer": "MOSIPNationalID",
                                                        "protocol": "OpenId4VCI",
                                                        "display": [
                                                            {
                                                                "name": "MOSIP National ID",
                                                                "logo": {
                                                                    "url": "https://api.collab.mosip.net/inji/mosip-logo.png",
                                                                    "alt_text": "MOSIP-Logo"
                                                                },
                                                                "title": "Download via LIC",
                                                                "description": "Enter your policy number to download your card.",
                                                                "language": "en"
                                                            }
                                                        ],
                                                        "client_id": "3yz7-j3xRzU3SODdoNgSGvO_cD8UijH3AIWRDAg1x-M",
                                                        ".well-known": "http://localhost:8088/.well-known/openid-credential-issuer"
                                                    }
                                                ]
                                            }
                                        },
                                        "errors": null
                                    },
                                    "Failure response": {
                                        "value": {
                                            "id": null,
                                            "version": "v1",
                                            "str": null,
                                            "responsetime": "2022-10-31T05:07:34.789Z",
                                            "metadata": null,
                                            "response": null,
                                            "errors": [
                                                {
                                                    "errorCode": "RESIDENT-APP-26",
                                                    "errorMessage": "Api not accessible failure"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}